Exercícios

+++ Capítulo 4/exercicio_4.7.9.c
@@ -5,9 +5,10 @@
 int main(){
 
 
-    float h;
+    double h;
     int p;
-    scanf("%f\t%d", &h, &p);
+    scanf("%lf\t%d", &h, &p);
+
 
     if (h<1.20 && p<60){
 
@@ -18,17 +19,23 @@ int main(){
     }
     else if (h<1.20 && p>90){
         printf("altura = %f\tpeso = %d\tclassificao = G\n", h,p);
-    }else if ((h>=1.20 && h<=1.70) && (p<60)){
+    }
+    else if ((h>=1.20 && h<=1.70) && (p<60)){
         printf("altura = %f\tpeso = %d\tclassificao = B\n", h,p);
-    }else if ((h>=1.20 && h<=1.70) && (p>=60&&p<=90)){
+    }
+    else if ((h>=1.20 && h<=1.70) && (p>=60&&p<=90)){
         printf("altura = %f\tpeso = %d\tclassificao = E\n", h,p);
-    }else if ((h>=1.20 && h<=1.70) && p>90){
+    }
+    else if ((h>=1.20 && h<=1.70) && p>90){
         printf("altura = %f\tpeso = %d\tclassificao = H\n", h,p);
-    }else if (h>1.70 && p<60){
+    }
+    else if (h > 1.70 && p < 60){
         printf("altura = %f\tpeso = %d\tclassificao = C\n", h,p);
-    }else if (h>1.70 && (p>=60 && p<=90)){
+    }
+    else if (h>1.70 && (p>=60 && p<=90)){
         printf("altura = %f\tpeso = %d\tclassificao = F\n", h,p);
-    }else if (h>1.70 && p>90){
+    }
+    else if (h>1.70 && p>90){
         printf("altura = %f\tpeso = %d\tclassificao = I\n", h,p);
         }
